Midievents release
I've created the "utils" directory to put all the necessary program
utilities.

All binaries from the programs which the source code is available in
this directory will be put on the bin/ (will be created automatically
when building the programs).

The Midievents program extract the timestamps for the notes played in
all tracks. You can specify the minimum time required between the notes
to consider them as different note events.