Flexible project configuration
To be able to easily configure the debug adapter now the OpenOCD flags
are stored in a separated file named ".prjconf.mk". This file is
generated by the config script (invoked when the file doesn't exists or
explicitly by executing "make config".

This can be expanded for including build configurations and other
customization in the future.