Microphone audio acquisition improved
The sound quality is acceptable now. In this commit I use a FIR
decimation filter to properly convert from PDM to PCM. The performance
is bad (almost all time  of the CPU is dedicated to the PDM to PCM
conversion), but I'll fixed it in the near future.

Changes introduced:
Decimation using a FIR filter (floating point);
Decimation factor increased from 64 to 128;
Sampling rate of the microphone increased from 1.024MHz to 2.048MHz.

As said the performance isn't any good, so I'll reduce the microphone
sampling rate to the original 1.024MHz and try to optimize the FIR
filter to work directly with the bit stream (no bit to float
conversion) in the next commits.

Happy New Year!