Audio double buffering and accumulator
To avoid race conditions, the audio buffer now uses a double buffering
scheme. Each buffer has 32000 bytes, so using a acquisition rate of
8KHz this represents 4s of audio.

A global variable called AudioAcc accumulates the abs value of all the
audio samples of one buffer. It is used to determine if there is
enough audio signal to proceed to the beat detection procedure.